backend/eventsync_api/requirements.txt
# dependencia de projeto
Django==5.0.4
djangorestframework==3.15.1

# dependencia de ambiente
python-decouple==3.8

# dependencia de login
djangorestframework-simplejwt==5.3.1

# dependencia de documentacao
drf-spectacular==0.27.2